refractor code, separate preset parameter

+++ windrecorder/record.py
@@ -84,9 +84,7 @@ def create_startup_shortcut(is_create=True):
             current_dir = os.getcwd()
             bat_path = os.path.join(current_dir, "start_record.bat")
             make_shortcut(bat_path, folder=startup_folder)
-            print(
-                "record: The shortcut has been created and added to the startup items"
-            )
+            print("record: The shortcut has been created and added to the startup items")
 
     else:
         # 移除快捷方式
@@ -101,9 +99,7 @@ def get_scale_screen_res_strategy(origin_width=1920, origin_height=1080):
     target_scale_width = origin_width
     target_scale_height = origin_height
 
-    if (
-        origin_height > 1500 and config.record_screen_enable_half_res_while_hidpi
-    ):  # 高分屏缩放至四分之一策略
+    if origin_height > 1500 and config.record_screen_enable_half_res_while_hidpi:  # 高分屏缩放至四分之一策略
         target_scale_width = int(origin_width / 2)
         target_scale_height = int(origin_height / 2)
 
@@ -111,7 +107,7 @@ def get_scale_screen_res_strategy(origin_width=1920, origin_height=1080):
 
 
 # 压缩视频分辨率到输入倍率
-def compress_video_resolution(video_path, scale_factor, config=config):
+def compress_video_resolution(video_path, scale_factor):
     scale_factor = float(scale_factor)
 
     # 获取视频的原始分辨率
@@ -119,48 +115,51 @@ def compress_video_resolution(video_path, scale_factor, config=config):
     output = subprocess.check_output(cmd, shell=True).decode("utf-8").strip()
     width, height = map(int, output.split(","))
 
-    # 计算目标分辨率
+    # 计算压缩视频的目标分辨率
     target_width = int(width * scale_factor)
     target_height = int(height * scale_factor)
 
     # 获取编码器和加速器
-    if config.compress_encoder == "x265":
-        encoder = "libx265"
-        crf_flag = "-crf "
-        if config.compress_accelerator == "qsv":
-            encoder = "hevc_qsv"
-            crf_flag = "-global_quality:v "
-        elif config.compress_accelerator == "nvenc":
-            encoder = "hevc_nvenc"
-            crf_flag = "-cq "
-    elif config.compress_encoder == "x264":
-        encoder = "libx264"
-        crf_flag = "-crf "
-        if config.compress_accelerator == "qsv":
-            encoder = "h264_qsv"
-            crf_flag = "-global_quality:v "
-        elif config.compress_accelerator == "nvenc":
-            encoder = "h264_nvenc"
-            crf_flag = "-cq "
-    elif config.compress_encoder == "av1":
-        encoder = "libaom-av1"
-        crf_flag = "-cpu-used 5 -row-mt 1 -tile-columns 2 -tile-rows 2 -crf "
+    encoder_default = config.compress_preset["x264"]["cpu"]["encoder"]
+    crf_flag_default = config.compress_preset["x264"]["cpu"]["crf_flag"]
+    crf_default = 39
+    try:
+        encoder = config.compress_preset[config.compress_encoder][config.compress_accelerator]["encoder"]
+        crf_flag = config.compress_preset[config.compress_encoder][config.compress_accelerator]["crf_flag"]
+        crf = int(config.compress_quality)
+    except KeyError:
+        print("Fail to get video compress config correctly. Fallback to default preset.")
+        encoder = encoder_default
+        crf_flag = crf_flag_default
+        crf = crf_default
 
-    crf = int(config.compress_quality)
+    # 执行压缩流程
+    def encode_video(encoder=encoder, crf_flag=crf_flag, crf=crf):
+        # 处理压缩视频路径
+        if "-OCRED" in os.path.basename(video_path):
+            output_newname = os.path.basename(video_path).replace("-OCRED", "-COMPRESS-OCRED")
+        else:  # 其他用途下的压缩用（如测试）
+            output_newname = "compressed_" + encoder + "_" + str(crf) + "_" + os.path.basename(video_path)
+        output_path = os.path.join(os.path.dirname(video_path), output_newname)
 
-    # 压缩视频分辨率
-    # output_path = f'compressed_{target_width}x{target_height}.mp4'
-    output_newname = os.path.basename(video_path).replace("-OCRED", "-COMPRESS-OCRED")
-    # output_newname = os.path.basename(video_path).replace(
-    # "-OCRED", f"-COMPRESS-OCRED-{encoder}-{crf}"
-    # )
-    output_path = os.path.join(os.path.dirname(video_path), output_newname)
+        # 如果输出目的已存在，将其移至回收站
+        if os.path.exists(output_path):
+            send2trash(output_path)
+
+        cmd = f"ffmpeg -i {video_path} -vf scale={target_width}:{target_height} -c:v {encoder} {crf_flag} {crf} {output_path}"
+        print(cmd)
+        subprocess.call(cmd, shell=True)
+        return output_path
+
+    # 如果系统不支持编码、导致输出的文件不正常或无输出，fallback 到默认参数
+    output_path = encode_video()
     if os.path.exists(output_path):
-        send2trash(output_path)
-
-    cmd = f"ffmpeg -i {video_path} -vf scale={target_width}:{target_height} -c:v {encoder} {crf_flag}{crf} {output_path}"
-
-    print(cmd)
-    subprocess.call(cmd, shell=True)
+        if os.stat(output_path).st_size < 1024:
+            print("Parameter not supported, fallback to default setting.")
+            send2trash(output_path)  # 清理空文件
+            output_path = encode_video(encoder=encoder_default, crf_flag=crf_flag_default, crf=crf_default)
+    else:
+        print("Parameter not supported, fallback to default setting.")
+        output_path = encode_video(encoder=encoder_default, crf_flag=crf_flag_default, crf=crf_default)
 
     return output_path
